MVP goal met! Photo Search feature:
As a user types in the search bar, the `onChange` event handler updates the value of `userInput` in the `SearchForPhotos` component state.
The `onChange` event handler also generates an (unencoded) query string for the API request, then uses the `encodeURI` function to escape any special characters in the user input.
When the submit button is clicked, the `setNextUrl` method from the `Gallery` component updates the value of `nextUrl` in state, then submits a new async API request.